Updates to scope model and view
I'm testing out one possible approach to this kind of models that would make it able to be used with the audio signals stemming from any stereo audio module. This is done by embedding a separate module (model + view) within the model. The embedded module handles parameters particular to configurating a scope~ GUI object, and enables the top level view to become a view of any audio module that use a stereo pair of jcom.out~.