Updating sur.aux~ for 0.6:
- Removing the global audio/gain and audio/mute parameters, these seem irrelevant and distracting in this context
- Changing internal gain unit to dB, seems more appropriate than midi values.
- Using parameterArray and messageArray objects
- aux.[n]/mute is now coupled to aux.[n]/gain: When the latter change we assume that the aux bus is to be used (so unmated), and when gain is set to -70 dB the bus is automatically muted.

NOTE: This module reveals a potential bug in AudioGraph: When aux.1 is muted, aux.2 and aux.3 turns silent. If aux.2 is muted, aux.3 is silenced. So it seems that muting of AudioGraph objects might influence the audio graph in unintended ways. This should be reviewed at the upcoming BEK workshop.